+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/CommitResponse.java
@@ -19,6 +19,7 @@ package com.google.cloud.spanner;
 import com.google.cloud.Timestamp;
 import com.google.common.base.Preconditions;
 import java.util.Objects;
+import javax.annotation.Nullable;
 
 /** Represents a response from a commit operation. */
 public class CommitResponse {
@@ -39,6 +40,18 @@ public class CommitResponse {
   /** Returns a {@link Timestamp} representing the commit time of the transaction. */
   public Timestamp getCommitTimestamp() {
     return Timestamp.fromProto(proto.getCommitTimestamp());
+  }
+
+  /**
+   * Returns a {@link Timestamp} representing the timestamp at which all reads in the transaction
+   * ran at, if the transaction ran at repeatable read isolation in internal test environments, and
+   * otherwise returns null.
+   */
+  public @Nullable Timestamp getSnapshotTimestamp() {
+    if (proto.getSnapshotTimestamp() == com.google.protobuf.Timestamp.getDefaultInstance()) {
+      return null;
+    }
+    return Timestamp.fromProto(proto.getSnapshotTimestamp());
   }
 
   /**