+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/connection/it/ITRetryDmlAsPartitionedDmlTest.java
@@ -86,6 +86,10 @@ public class ITRetryDmlAsPartitionedDmlTest extends ITAbstractSpannerTest {
 
   @Test
   public void testDmlFailsIfMutationLimitExceeded() {
+    // TODO(sakthivelmani) - Re-enable once b/422916293 is resolved
+    assumeFalse(
+        "Skipping the test due to a known bug b/422916293",
+        env.getTestHelper().getOptions().isEnableDirectAccess());
     try (Connection connection = createConnection()) {
       connection.setAutocommit(true);
       assertThrows(
@@ -98,6 +102,10 @@ public class ITRetryDmlAsPartitionedDmlTest extends ITAbstractSpannerTest {
 
   @Test
   public void testRetryDmlAsPartitionedDml() throws Exception {
+    // TODO(sakthivelmani) - Re-enable once b/422916293 is resolved
+    assumeFalse(
+        "Skipping the test due to a known bug b/422916293",
+        env.getTestHelper().getOptions().isEnableDirectAccess());
     try (Connection connection = createConnection()) {
       connection.setAutocommit(true);
       connection.setAutocommitDmlMode(
@@ -138,6 +146,10 @@ public class ITRetryDmlAsPartitionedDmlTest extends ITAbstractSpannerTest {
 
   @Test
   public void testRetryDmlAsPartitionedDml_failsForLargeInserts() throws Exception {
+    // TODO(sakthivelmani) - Re-enable once b/422916293 is resolved
+    assumeFalse(
+        "Skipping the test due to a known bug b/422916293",
+        env.getTestHelper().getOptions().isEnableDirectAccess());
     try (Connection connection = createConnection()) {
       connection.setAutocommit(true);
       connection.setAutocommitDmlMode(