+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITWriteTest.java
@@ -1081,10 +1081,8 @@ public class ITWriteTest {
       write(baseInsert().set("StringValue").to(1.234).build());
       fail("Expected exception");
     } catch (SpannerException ex) {
-      if (env.getTestHelper()
-          .getOptions()
-          .getSessionPoolOptions()
-          .getUseMultiplexedSessionForRW()) {
+      if (env.getTestHelper().getOptions().getSessionPoolOptions().getUseMultiplexedSessionForRW()
+          && !isUsingEmulator()) {
         assertThat(ex.getErrorCode()).isEqualTo(ErrorCode.INVALID_ARGUMENT);
       } else {
         assertThat(ex.getErrorCode()).isEqualTo(ErrorCode.FAILED_PRECONDITION);